requirements.txt
numpy==1.21.0
scipy==1.7.1
matplotlib==3.4.3
Pillow==8.4.0
